Reorganize utils and components
