+++ plugins/woo-ai/src/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './productData';
+export * from './shuffleArray';
