+++ plugins/woo-ai/src/utils/productData.ts
