+++ plugins/woo-ai/src/utils/types.ts
